remove plugin helpers file
