+++ packages/datadog-plugin-http/test/integration-test/server.mjs
@@ -1,4 +1,14 @@
 import 'dd-trace/init.js'
-import * as pluginHelpers from './plugin-helpers.mjs'
+import http from 'http'
 
-pluginHelpers.onMessage(async () => {})
+const server = http.createServer(async (req, res) => {
+  try {
+    res.end('integration test response handler success')
+  } catch (err) {
+    res.statusCode = 500
+    res.end('integration test response handler failure')
+  }
+}).listen(0, () => {
+  const port = server.address().port
+  process.send({ port })
+})
